feat: compensate pagination things for multiple data sources

+++ src/lib/api_app/src/api/routes/testbed/productizers/job/job_models.rs
@@ -17,8 +17,8 @@ pub struct RequestLocation {
 }
 #[derive(Deserialize, Serialize, Debug)]
 pub struct RequestPaging {
-    pub limit: usize,
-    pub offset: usize,
+    pub limit: i32,
+    pub offset: i32,
 }
 
 //
